+++ src/tables/mod.rs
@@ -23,3 +23,4 @@ pub mod maxp;
 pub mod name;
 pub mod os2;
 pub mod post;
+pub mod sbix;